requirements.txt
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
fastapi==0.63.0
google-auth==1.27.0
google-auth-oauthlib==0.4.2
gspread==3.7.0
gunicorn==20.0.4
h11==0.12.0
idna==2.10
oauthlib==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8
pyTelegramBotAPI==3.7.6
python-dotenv==0.15.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
starlette==0.13.6
typing-extensions==3.7.4.3
urllib3==1.26.3
uvicorn==0.13.4
